Update the nfs_acl.x copyright notices
Greg suggests:

For the nfsacl.x file, the author should be updated to Oracle.
The original header should not be preserved (drop the separate comment block
for Copyright 1994..Sun Microsystems All Rights Reserved).
It only needs the copyright one time.

If we want to maintain the historical copyright header (not sure how that
overlaps with the IETF copyright) it should read:

Copyright © 1994, 2024, Oracle and/or its affiliates.