Remove descriptions of local file system ACL semantics
A recent conversation with Rick Macklem convinced me (after Rob also warned about it) that this document should not and cannot specify ACL semantics that already exist and are particular to local file systems. Therefore, weaken or remove language that tries to specify behaviors that should be left to the NFS server implementation or the local file systems that it controls.