Address review comment from Pali Rohar
Pali notes:

> 3.5.4.3. Relationship Between ACLs and Other File Attributes
> When an ACL is present on a file, changing the file's owner (say, via the SETATTR operation) may alter the server's interpretation of any ACE that targets @OWNER.
> When an ACL is present on a file, changing the file's group (say, via the SETATTR operation) may alter the server's interpretation of any ACE that targets @GROUP.

There is missing definition what is the "@OWNER" and "@GROUP". In the
whole draft these terms are not defined.

In NFS4 there is defined "OWNER@" and "GROUP@" (with @ at the end). And
important is that NFS4 GROUP@ is not NFS3 NA_GROUP_OBJ as defined in
this draft. If I understood it correctly then following applies:
NFS3 NA_GROUP_OBJ = NFS4 GROUP@ - NFS4 OWNER@

I guess that instead of "@OWNER" and "@GROUP" there should be
NA_USER_OBJ and NA_GROUP_OBJ.