Rename u_int
The base integer types as described in RFC 4506 do not include a "u_int" type. It's a platform- or library-specific alias (often a BSD- or BSD-derived extension) for an unsigned integer type, not part of the standard C types described. Since this isn't either a standardized RPC or C type, replace "u_int" with "unsigned integer" throughout.